refactor: reorganize silent import flow
